+++ NEWS.md
@@ -11,3 +11,4 @@
 - Parse Spectrum births in read_specdp_demog_param().
 - Add births from Spectrum to specfp object.
 - For epp regions, distribute number of births based on 15-44 population size and GFR.
+- Implement age 15 entrant prevalence and ART coverage inputs.